Moved changes to base Compat CUP Vehicles
