+++ addons/compat_cup_vehicles/CfgVehicles.hpp
@@ -419,4 +419,13 @@ class CfgVehicles {
             "hitera_r16", "hitera_r17", "hitera_r18", "hitera_r19", "hitera_r20"
         };
     };
+    class Helicopter_Base_H;
+    class CUP_KA50_Base: Helicopter_Base_H {
+        class pilotCamera;
+    };
+    class CUP_KA50_Dynamic_Base: CUP_KA50_Base {
+        class pilotCamera: pilotCamera {
+            EGVAR(missileguidance,usePilotCameraForTargeting) = 1;
+        };
+    };
 };
