--- addons/compat_cup_vehicles/compat_cup_vehicles_missileguidance/config.cpp
@@ -1,21 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches {
-    class SUBADDON {
-        name = COMPONENT_NAME;
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {
-            "ace_missileguidance",
-            "CUP_AirVehicles_Ka50"
-        };
-        skipWhenMissingDependencies = 1;
-        author = ECSTRING(common,ACETeam);
-        url = ECSTRING(main,URL);
-        VERSION_CONFIG;
-        addonRootClass = QUOTE(ADDON);
-    };
-};
-
-#include "CfgVehicles.hpp"
