--- addons/compat_cup_vehicles/compat_cup_vehicles_missileguidance/script_component.hpp
@@ -1,3 +1,0 @@
-#define SUBCOMPONENT missileguidance
-#define SUBCOMPONENT_BEAUTIFIED Missile Guidance
-#include "..\script_component.hpp"
